hellowealth-backend/requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
coreapi==2.3.3
Django==3.1.6
djangorestframework==3.12.2
django-allauth==0.44.0
django-cors-headers==3.7.0
dj-rest-auth==2.1.3
djangorestframework-simplejwt==4.6.0
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytz==2021.1
setuptools==53.0.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
appdirs==1.4.4
botocore==1.20.9
cement==3.0.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.6.0
distlib==0.3.1
filelock==3.0.12
future==0.18.2
idna==2.10
install==1.3.4
jmespath==0.10.0
oauthlib==3.1.0
pathspec==0.8.1
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
semantic-version==2.8.5
termcolor==1.1.0
urllib3==1.26.3
wcwidth==0.2.5
drf_yasg
pymongo
dnspython